Refactor analytics into a single package
